LayoutTests/fast/viewport/viewport-24.html

<head>
    <title>Viewport META with "maximum-scale=3, minimum-scale=4".</title>
    <meta name="viewport" content="maximum-scale=3, minimum-scale=4">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(160, 320, 480, 320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
