Safari-611.1.10/LayoutTests/fast/viewport/viewport-24-expected.txt

ALERT: viewport size 980x1078 scale 4 with limits [4, 4] and userScalable true
